countdown & daily challenge refactor
