+++ src/components/Countdown.tsx
@@ -1,22 +1,28 @@
 'use client';
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { memo, useEffect, useMemo, useState } from 'react';
+import { setHours, startOfToday, startOfTomorrow } from 'date-fns';
 
 import { Stack } from '@mui/material';
 import accurateInterval from 'accurate-interval';
 
 import Text from '@/components/Text';
 
-export const CountdownTimer = ({
-  fontSize = '100px',
-  color = 'black',
-}: {
-  fontSize?: string;
-  color?: string;
-}) => {
-  const targetDate = new Date('December 10, 2023 17:00:00').getTime();
+function addLeadingZero(num: number) {
+  return num < 10 ? `0${num}` : num;
+}
+
+export const CountdownTimer = memo(() => {
+  // targetDate is everyday at 5PM except the last day of the game
+  const targetDate = setHours(startOfToday(), 17).getTime();
+
   const calculateInitialCount = () => {
     const currentTime = new Date().getTime();
-    const timeDifference = targetDate - currentTime;
+    let timeDifference = targetDate - currentTime;
+    if (timeDifference < 0) {
+      // need to get tmr 5pm
+      timeDifference = setHours(startOfTomorrow(), 17).getTime();
+    }
+
     return Math.floor(timeDifference / 1000); // Convert to seconds
   };
 
@@ -36,19 +42,28 @@ export const CountdownTimer = ({
 
   // Function to format the countdown
   const formatCountdown = useMemo(() => {
-    const days = Math.floor(count / (3600 * 24));
     const hours = Math.floor((count % (3600 * 24)) / 3600);
     const minutes = Math.floor((count % 3600) / 60);
     const seconds = count % 60;
 
-    return `${days}d ${hours}h ${minutes}m ${seconds}s`;
+    return `${addLeadingZero(hours)}: ${addLeadingZero(minutes)}: ${addLeadingZero(seconds)}`;
   }, [count]);
 
   return (
-    <Stack width="100%" gap={8} textAlign="center">
-      <Text fontSize={fontSize} useMonoFont color={color}>
+    <Stack
+      direction="column"
+      width="100%"
+      gap={1}
+      borderRadius="12px"
+      p={2}
+      sx={{ backgroundColor: 'white' }}
+    >
+      <Text variant="h5" useMonoFont>
         {formatCountdown}
       </Text>
+      <Text variant="body1">Time remaining</Text>
     </Stack>
   );
-};
+});
+
+CountdownTimer.displayName = 'CountdownTimer';
